--- src/components/First.js
@@ -1,9 +1,0 @@
-import { Text } from "@chakra-ui/react";
-
-export default function First() {
-  return (
-    <Text fontSize="xl" fontWeight="bold" color="orange.600" mt="10%">
-      추천 도서
-    </Text>
-  );
-}
