+++ src/components/Recommend.js
@@ -4,23 +4,28 @@ import { Flex, Grid, Box, Text, Spinner, Image } from "@chakra-ui/react";
 import bookCover from "../../public/img/book-cover.png";
 
 export default function Recommend() {
-  const [bookObj, setBookObj] = useState([]);
+  const [related, setRelated] = useState([]);
   const [loading, setLoading] = useState(false);
-  const author = "황순원";
+
+  function randomItem(a) {
+    return a[Math.floor(Math.random() * a.length)];
+  }
 
   useEffect(() => {
     const book = async () => {
       setLoading(true);
-      if (author) {
+      if (randomItem(recommendKeyword)) {
         const response = await axios.get(
-          `https://www.googleapis.com/books/v1/volumes?q=${author}`
+          `https://www.googleapis.com/books/v1/volumes?q=${randomItem(
+            recommendKeyword
+          )}`
         );
-        setBookObj(response.data.items);
+        setRelated(response.data.items);
       }
       setLoading(false);
     };
     book();
-    console.log(bookObj);
+    console.log(related);
   }, []);
 
   if (loading)
@@ -31,7 +36,7 @@ export default function Recommend() {
     );
   return (
     <Grid templateColumns="repeat(4,1fr)" m={10}>
-      {bookObj.map((value, index) => (
+      {related.map((value, index) => (
         <Box key={index} position="relative" visibility="">
           {value.volumeInfo.imageLinks ? (
             <Image
@@ -66,3 +71,28 @@ export default function Recommend() {
     </Grid>
   );
 }
+
+const recommendKeyword = [
+  "시",
+  "소설",
+  "유아",
+  "에세이",
+  "경제",
+  "여행",
+  "사회",
+  "컴퓨터",
+  "뷰티",
+  "건강",
+  "전공",
+  "만화",
+  "역사",
+  "문화",
+  "경영",
+  "세계",
+  "기술",
+  "미술",
+  "음악",
+  "체육",
+  "취미",
+  "외국어",
+];
